Move some test-related logic into the forcing module so it can be re-used
